quant-bot/requirements.txt
# Core
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
optuna>=3.3.0  # Hyperparameter optimization

# Deep Learning (optional)
torch>=2.0.0
# tensorflow>=2.13.0

# Technical Analysis
ta-lib>=0.4.28
pandas-ta>=0.3.14b0

# Data
yfinance>=0.2.28
ccxt>=4.0.0
pyarrow>=13.0.0  # Parquet support

# API & Web
requests>=2.31.0
websocket-client>=1.6.0
aiohttp>=3.8.5
fastapi>=0.100.0
uvicorn>=0.23.0
beautifulsoup4>=4.12.0

# NLP & Sentiment
transformers>=4.30.0
ftfy>=6.1.1
regex>=2023.6.3

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Jupyter
jupyter>=1.0.0
ipywidgets>=8.1.0
nbformat>=5.9.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
tqdm>=4.66.0
loguru>=0.7.0

# Type hints
mypy>=1.5.0
types-requests>=2.31.0

# Linting
black>=23.9.0
ruff>=0.0.290
isort>=5.12.0

# Numba for JIT compilation
numba>=0.58.0

# Statistical analysis
statsmodels>=0.14.0
arch>=6.1.0  # GARCH models

# Vector DB & Memory
chromadb>=0.4.0
sentence-transformers>=2.2.0
scikit-optimize>=0.9.0
